Fix metadata_cubes test with newer gdal versions
The test was failing due to trying to read from a raster that had not
yet been flushed. The same dataset was opened for writing and for
reading simultaneously, which is not safe in general since changes
written to the raster may be cached until the raster object is
destroyed.

The fix is to ensure that the original raster object goes out of scope
(so that changes are guaranteed to be flushed) before re-opening it for
reading. In other cases, we instead simply read from the original raster
object to instead of re-opening it for reading.

(cherry picked from commit ace31deef7acbe2fe258e6bbf0df25eff5edc5bc)