Update contributing guide (#2190)
Update the contributing guide to highlight best practices employed to
ensure that we are able to safely accept external contributions.

The important bits that I think we should highlight are:

* Only isce3 maintainers are able to directly write to the repository
  (either by creating new branches or modifying existing branches)
* All PRs must be approved by >=2 core team members before they can be
  merged by a maintainer
* CI tests, security scanning, and linter tools provide automated checks
  that ensure contributions are high quality and conform to our code
  standards

One point that's not mentioned here but worth noting for the RITM
reviewers is that we plan to enable Dependabot alerts and updates for
any GitHub Actions used. I don't think it makes sense to shoehorn that
into the contributing guide, but it's a good policy that JPL should
consider recommending for public repositories that use GitHub Actions.